Update dmx values (#30)
* Update define here

* Add some call here

* Update GDTFManager.cpp

* Refactor

* Refactor DmxAddress

* Updated dependency build order for lib and testapp

* Fix on the mac

* Fix the value here

* Put this in here as well

* Add Higlight Unit Test

* Make this clear

* Fix this here

* Win fix, lets see if other platforms work too

* Fix all platforms (hopefully, look at travis)

* Fix comment

* Fix source Definition like Header before

+++ src/Include/GDTFPrefix.h
@@ -9,8 +9,12 @@ namespace VectorworksMVR
 {
 	namespace GdtfDefines
 	{
-		typedef size_t		DMXAddress;
-		typedef size_t		DmxValue;
+        // This is used for DMX Channel defining. So values from 1...512 will be stored
+		typedef Uint16		DMXAddress; 
+		
+        // This is used for a DMX Value defining. GDTF currently supports of to 32bit values, so the maximum value will be 2^32 - 1
+        // We always store this in a 64bit values, to make this more simple. 
+        typedef Uint64		DmxValue;    
 		
 		enum EGdtfObjectType
 		{
